Rename validation -> analysis
